SRC/picmoney/backend/requirements.txt
# Web
fastapi==0.114.0
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1

# Data (wheels estáveis para Py 3.11)
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.1
openpyxl==3.1.5

# ORM e validação
SQLAlchemy==2.0.35
pydantic==2.8.2
pydantic-settings==2.4.0

# Utilitários
tabulate==0.9.0
chardet==5.2.0
Unidecode==1.3.8

flask-jwt-extended>=4.6.0
passlib[bcrypt]>=1.7.4
